backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
openai==1.3.0
pinecone-client==3.0.0
google-generativeai==0.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
pandas==2.1.3
